HBAI-330 Make storage into a class.
