+++ backends/embeddings/route.py
@@ -1,144 +1,13 @@
 import os
 import json
 from datetime import datetime, timezone
-from typing import Any, List
 from core import classes, common
 from fastapi import APIRouter, Request, Depends, File, BackgroundTasks, UploadFile
-from . import storage, file_parsers, main
+from embeddings.vector_storage import Vector_Storage
+from . import file_parsers
+from embeddings.embedder import modify_document, delete_sources
 
 router = APIRouter()
-
-###############
-### Methods ###
-###############
-
-
-# Given source(s), delete all associated document chunks, metadata and files
-def delete_sources(
-    app: Any, collection_name: str, sources: List[classes.SourceMetadata]
-):
-    db = storage.get_vector_db_client(app)
-    collection = db.get_collection(name=collection_name)
-    vector_index = main.load_embedding(app, collection_name)
-    # Delete each source chunk and parsed file
-    for source in sources:
-        chunk_ids = source.get("chunkIds")
-        # Delete all chunks
-        storage.delete_chunks(
-            collection=collection,
-            vector_index=vector_index,
-            chunk_ids=chunk_ids,
-        )
-        # Delete associated files
-        storage.delete_source_files(source)
-    # Update collection metadata.sources to remove this source
-    storage.update_collection_sources(
-        collection=collection,
-        sources=sources,
-        mode="delete",
-    )
-
-
-async def modify_document(
-    app: Any,
-    form: classes.EmbedDocumentRequest,
-    file: UploadFile,
-    background_tasks: BackgroundTasks,
-    is_update: bool = False,
-):
-    document_name = form.documentName
-    prev_document_id = form.documentId
-    source_name = form.documentName
-    collection_name = form.collectionName
-    description = form.description
-    tags = common.parse_valid_tags(form.tags)
-    url_path = form.urlPath
-    local_file_path = form.filePath
-    text_input = form.textInput
-    chunk_size = form.chunkSize
-    chunk_overlap = form.chunkOverlap
-    chunk_strategy = form.chunkStrategy
-    parsing_method = form.parsingMethod
-    new_document_id = file_parsers.create_parsed_id(collection_name=collection_name)
-    if is_update:
-        source_id = prev_document_id
-    else:
-        source_id = new_document_id
-    form_data = {
-        "collection_name": collection_name,
-        "document_name": document_name,
-        "document_id": source_id,
-        "description": description,
-        "tags": tags,
-        "embedder": app.state.embed_model,
-        "chunk_size": chunk_size,
-        "chunk_overlap": chunk_overlap,
-        "chunk_strategy": chunk_strategy,
-        "parsing_method": parsing_method,
-    }
-    # Verify input values
-    if (
-        file == None  # file from client
-        and url_path == ""  # file on web
-        and text_input == ""  # text input from client
-        and local_file_path == ""  # file on server disk
-    ):
-        raise Exception("Please supply a file upload, file path, url or text.")
-    if not collection_name or collection_name == "undefined" or not source_name:
-        raise Exception("Please supply a collection name and/or memory name.")
-    if is_update and not prev_document_id:
-        raise Exception("Please supply a document id.")
-    if not document_name:
-        raise Exception("Please supply a document name.")
-    if not source_id:
-        raise Exception("Server error, id misconfigured.")
-    if not common.check_valid_id(source_name):
-        raise Exception(
-            "Invalid memory name. No '--', uppercase, spaces or special chars allowed."
-        )
-    if tags == None:
-        raise Exception("Invalid value for 'tags' input.")
-    # If updating, Remove specified source(s) from database
-    if is_update:
-        collection = storage.get_collection(app, name=collection_name)
-        sources_to_delete = storage.get_sources_from_ids(
-            collection=collection, source_ids=[prev_document_id]
-        )
-        delete_sources(
-            app=app, collection_name=collection_name, sources=sources_to_delete
-        )
-    # Write uploaded file to disk temporarily
-    input_file = await file_parsers.copy_file_to_disk(
-        app=app,
-        url_path=url_path,
-        file_path=local_file_path,
-        text_input=text_input,
-        file=file,
-        id=source_id,
-    )
-    path_to_parsed_file = input_file.get("path_to_file")
-    # Read in files and create index nodes
-    nodes = await main.create_index_nodes(
-        app=app,
-        input_file=input_file,
-        form=form_data,
-    )
-    # Create embeddings
-    # @TODO Note that you must NOT perform CPU intensive computations in the background_tasks of the app,
-    # because it runs in the same async event loop that serves the requests and it will stall your app.
-    # Instead submit them to a thread pool or a process pool.
-    background_tasks.add_task(
-        main.create_new_embedding,
-        nodes=nodes,
-        form=form_data,
-        app=app,
-    )
-    return path_to_parsed_file
-
-
-##############
-### Routes ###
-##############
 
 
 @router.get("/addCollection")
@@ -167,8 +36,8 @@ def create_memory_collection(
             "description": form.description,
             "sources": json.dumps([]),
         }
-        db_client = storage.get_vector_db_client(app)
-        db_client.create_collection(
+        vector_storage = Vector_Storage(app=app)
+        vector_storage.db_client.create_collection(
             name=collection_name,
             metadata=metadata,
         )
@@ -279,7 +148,8 @@ def get_all_collections(
     app = request.app
 
     try:
-        collections = storage.get_all_collections(app=app)
+        vector_storage = Vector_Storage(app=app)
+        collections = vector_storage.get_all_collections()
 
         return {
             "success": True,
@@ -305,7 +175,8 @@ def get_collection(
 
     try:
         name = props.id
-        collection = storage.get_collection(app=app, name=name)
+        vector_storage = Vector_Storage(app=app)
+        collection = vector_storage.get_collection(name=name)
 
         return {
             "success": True,
@@ -330,8 +201,9 @@ def get_chunks(request: Request, params: classes.GetDocumentChunksRequest):
     app = request.app
 
     try:
-        chunks = storage.get_source_chunks(
-            app, collection_name=collection_id, source_id=document_id
+        vector_storage = Vector_Storage(app=app)
+        chunks = vector_storage.get_source_chunks(
+            collection_name=collection_id, source_id=document_id
         )
         if chunks != None:
             num_chunks = len(chunks)
@@ -385,8 +257,9 @@ def delete_document_sources(
         source_ids = params.document_ids
         num_documents = len(source_ids)
         # Find source data
-        collection = storage.get_collection(app, name=collection_name)
-        sources_to_delete = storage.get_sources_from_ids(
+        vector_storage = Vector_Storage(app=app)
+        collection = vector_storage.get_collection(name=collection_name)
+        sources_to_delete = vector_storage.get_sources_from_ids(
             collection=collection, source_ids=source_ids
         )
         # Remove specified source(s)
@@ -416,16 +289,17 @@ def delete_collection(
 
     try:
         collection_id = params.collection_id
-        db = storage.get_vector_db_client(app)
+        vector_storage = Vector_Storage(app=app)
+        db = vector_storage.db_client
         collection = db.get_collection(collection_id)
         # Remove all the sources in this collection
-        sources = storage.get_collection_sources(collection)
+        sources = vector_storage.get_collection_sources(collection)
         # Remove all associated source files
         delete_sources(app=app, collection_name=collection_id, sources=sources)
         # Remove the collection
         db.delete_collection(name=collection_id)
         # Remove persisted vector index from disk
-        common.delete_vector_store(collection_id, storage.VECTOR_STORAGE_PATH)
+        common.delete_vector_store(collection_id, vector_storage.VECTOR_STORAGE_PATH)
         return {
             "success": True,
             "message": f"Removed collection [{collection_id}]",
@@ -447,12 +321,13 @@ def wipe_all_memories(
 
     try:
         # Delete all db values
-        db = storage.get_vector_db_client(app)
+        vector_storage = Vector_Storage(app=app)
+        db = vector_storage.db_client
         db.reset()
         # Delete all parsed files in /memories
         file_parsers.delete_all_files()
         # Remove all vector storage collections and folders
-        storage.delete_all_vector_storage()
+        vector_storage.delete_all_vector_storage()
         # Acknowledge success
         return {
             "success": True,
